Split the third-party CLI agents subpage out of the AI settings page into its own page
